chore: Move adapters internal code
