+++ packages/nuqs/src/adapters/custom.ts
@@ -1,11 +1,11 @@
 export { renderQueryString } from '../url-encoding'
+export {
+  createAdapterProvider as unstable_createAdapterProvider,
+  type AdapterContext as unstable_AdapterContext
+} from './lib/context'
 export type {
   AdapterInterface as unstable_AdapterInterface,
   AdapterOptions as unstable_AdapterOptions,
   UpdateUrlFunction as unstable_UpdateUrlFunction,
   UseAdapterHook as unstable_UseAdapterHook
-} from './defs'
-export {
-  createAdapterProvider as unstable_createAdapterProvider,
-  type AdapterContext as unstable_AdapterContext
-} from './internal.context'
+} from './lib/defs'
