+++ packages/nuqs/src/adapters/next/pages.ts
@@ -1,4 +1,4 @@
-import { createAdapterProvider } from '../internal.context'
+import { createAdapterProvider } from '../lib/context'
 import { useNuqsNextPagesRouterAdapter } from './impl.pages'
 
 export const NuqsAdapter = createAdapterProvider(useNuqsNextPagesRouterAdapter)
