+++ packages/nuqs/src/adapters/react-router.ts
@@ -3,8 +3,8 @@
 
 import { useNavigate, useSearchParams } from 'react-router-dom'
 import { renderQueryString } from '../url-encoding'
-import type { AdapterOptions } from './defs'
-import { createAdapterProvider } from './internal.context'
+import { createAdapterProvider } from './lib/context'
+import type { AdapterOptions } from './lib/defs'
 
 function useNuqsReactRouterV6Adapter() {
   const navigate = useNavigate()
